+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 -   Added onKeyPress / onFocus / onBlur callbacks for all components with the option to return false to prevent the default action
 -   Fixed issue with addChild if the child has already been added
+-   Added position type for relative/absolute positioning
 
 # 0.3.0